+++ src/App.vue
@@ -1,12 +1,11 @@
 <script setup>
-  import HelloWorld from "./components/HelloWorld.vue";
   import Navbar from "./components/Navbar.vue";
 </script>
 
 <template>
   <Navbar />
-  <img alt="Vue logo" src="./assets/logo.png" />
-  <HelloWorld msg="Hello, welcome to my personal website!" />
+  <img alt="My logo" src="./assets/logo.png" />
+  <h1>Hello, welcome to my personal website!</h1>
   <a href="https://github.com/NoteNecron/Note/">Source Code!</a>
 </template>